reacomodado promo y añadido ruta
