--- promo.html
@@ -1,71 +1,0 @@
-<%- include('partials/_header') -%>
-<body>
-  <section id="new">
-      <h3 style="font-size: 50px; font-weight: 500; text-align: center;">PROMOCIONES</h3>
-      <div class="accordion" id="accordionExample">
-<div class="card">
-  <div class="card-header" id="headingOne">
-    <h2 class="mb-0">
-      <button class="btn btn-link" type="button" data-toggle="collapse" data-target="#collapseOne" aria-expanded="true" aria-controls="collapseOne" style="color: #3f3f40; font-size: 25px; font-weight: 400;">
-        Artículo de Regalo
-      </button>
-    </h2>
-  </div>
-
-  <div id="collapseOne" class="collapse show" aria-labelledby="headingOne" data-parent="#accordionExample">
-    <div class="card-body">
-      Obtén artículos de regalo con el siguiente código
-        <a  href="regaloof.html">Ver más</a>
-    </div>
-  </div>
-</div>
-<div class="card">
-  <div class="card-header" id="headingTwo">
-    <h2 class="mb-0">
-      <button class="btn btn-link collapsed" type="button" data-toggle="collapse" data-target="#collapseTwo" aria-expanded="false" aria-controls="collapseTwo" style="color: #3f3f40; font-size: 25px; font-weight: 400;">
-        Promo Parejas
-      </button>
-    </h2>
-  </div>
-  <div id="collapseTwo" class="collapse" aria-labelledby="headingTwo" data-parent="#accordionExample">
-    <div class="card-body">
-     Tú y tu pareja ahora pueden combinar atuendos
-        <a  href="regaloof.html">Ver más</a>
-    </div>
-  </div>
-</div>
-<div class="card">
-  <div class="card-header" id="headingThree">
-    <h2 class="mb-0">
-      <button class="btn btn-link collapsed" type="button" data-toggle="collapse" data-target="#collapseThree" aria-expanded="false" aria-controls="collapseThree" style="color: #3f3f40; font-size: 25px; font-weight: 400;">
-        Aeropostale en Rebajas
-      </button>
-    </h2>
-  </div>
-  <div id="collapseThree" class="collapse" aria-labelledby="headingThree" data-parent="#accordionExample">
-    <div class="card-body">
-      Artículos a mitad de precio
-        <a  href="regaloof.html">Ver más</a>
-    </div>
-  </div>
-</div>
-          <div class="card">
-  <div class="card-header" id="headingThree">
-    <h2 class="mb-0">
-      <button class="btn btn-link collapsed" type="button" data-toggle="collapse" data-target="#collapseFour" aria-expanded="false" aria-controls="collapseFour" style="color: #3f3f40; font-size: 25px; font-weight: 400;">
-        Código especial
-      </button>
-    </h2>
-  </div>
-  <div id="collapseFour" class="collapse" aria-labelledby="headingFour" data-parent="#accordionExample">
-    <div class="card-body">
-      Por asuntos de COVID-19, les damos este código de descuento
-        <a  href="regaloof.html">Ver más</a>
-    </div>
-  </div>
-</div>
-</div>
-  </section>
-      </body>
-<%- include('partials/_footer') -%>
-</html>
