--- promoController.js
@@ -1,9 +1,0 @@
-const controller = {};
-
-controller.list = (req, res) =>{
-    res.render('promo.html', {
-        title: 'StyleU'
-    });
-
-};
-module.exports = controller;
