+++ src/controllers/promoController.js
@@ -1,0 +1,9 @@
+const controller = {};
+
+controller.list = (req, res) =>{
+    res.render('promo.html', {
+        title: 'StyleU'
+    });
+
+};
+module.exports = controller;
